+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2024-03-02
+
+- Replace deadpool-sqlite with a connector based on tokio-rusqlite.
+- Rework error system.
+
 ## [0.3.3] - 2023-10-08
 
 ## Changed